+++ packages/tendermint-rpc/src/tendermint37/adaptor/index.ts
@@ -1,13 +1,2 @@
-import { hashBlock, hashTx } from "../hasher";
-import { Params } from "./requests";
-import { Responses } from "./responses";
-import { Adaptor } from "./types";
-
-export { Decoder, Encoder, Params, Responses } from "./types";
-
-export const adaptor37: Adaptor = {
-  params: Params,
-  responses: Responses,
-  hashTx: hashTx,
-  hashBlock: hashBlock,
-};
+export { Params } from "./requests";
+export { Responses } from "./responses";