+++ scripts/tendermint/all_stop.sh
@@ -4,7 +4,7 @@ command -v shellcheck >/dev/null && shellcheck "$0"
 
 SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
-for KEY in 34 37; do
+for KEY in 34 37 38; do
   export TENDERMINT_NAME="tendermint-$KEY"
 
   echo "Stopping $TENDERMINT_NAME ..."